modularize desktop shell data hook
